% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMELDThree.R
\name{calcMELDThree}
\alias{calcMELDThree}
\title{Calculate MELD-3.0 score}
\usage{
calcMELDThree(sex, creatinine, inr, bilirubin, sodium, albumin)
}
\arguments{
\item{creatinine}{a serum biomarker that reflects the kidney function. Unit:
mg/dl}

\item{inr}{International Normalized Ratio is a blood clotting test that is a
measure how quickly your blood clots and basically how thin your blood is.}

\item{bilirubin}{a biomarker that measures the ability of the liver to
process this byproduct from the breakdown of old red blood cells.  Elevated
levels in the blood (hyperbilirubinemia) indicate poor liver function
unit: mg/dl}

\item{sodium}{serum sodium level. Unit: mEq/L}

\item{albumin}{A liver synthesized protein which low levels of can indicate
chronic liver disease or liver dysfunction. Unit: mg/dL}

\item{Sex}{a boolean value indicating whether the patient is female or not.
As females get extra points in the calculation to account for biases.}
}
\value{
meldThreeScore The MELD-3.0 score calculation based on the paper by
Kim et al.
}
\description{
A function that calculates the MELD-3.0 score per Kim et al.
}
\examples{
#Example 1
# Calculate the MELD-3.0 score for a female

calcMELDThree(True, 2.3, 1.9, 2, 145)

}
\references{
Kim WR, Mannalithara A, Heimbach JK, Kamath PS, Asrani SK, Biggins SW,
Wood NL, Gentry SE, Kwong AJ. MELD 3.0: The Model for End-Stage Liver Disease
Updated for the Modern Era. Gastroenterology. 2021 Dec;161(6):1887-1895.e4.
doi: 10.1053/j.gastro.2021.08.050. Epub 2021 Sep 3. PMID: 34481845;
PMCID: PMC8608337.

Kamath PS, Wiesner RH, Malinchoc M, Kremers W, Therneau TM, Kosberg CL,
D'Amico G, Dickson ER, Kim WR. A model to predict survival in patients with
end-stage liver disease. Hepatology. 2001 Feb;33(2):464-70.
doi: 10.1053/jhep.2001.22172. PMID: 11172350.
}
