% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMELDdf.R
\name{createMELDdf}
\alias{createMELDdf}
\title{Create a data frame with copmuted MELD, MELD-na and MELD3.0 scores for each
entry}
\usage{
createMELDdf(data)
}
\arguments{
\item{data}{Dataframe created by the prepareData function. Needs to contain
the following columns:
Patient_ID, Data, Albumin, Bilirubin, INR, Serum_Creatinine, Serum_Sodium,
Gender, Dialysis_In_Past_Week.}
}
\value{
MELDdf, a dataframe containing the following columns: Patient_ID,
Date, MELD, MELDna, MELD3_0.
It will contain the MELD, MELD-na and MELD-3.0 scores from the input
dataframe containing the Liver transplantation biomarkers and clinical data
}
\description{
This function uses the calcMELD, calcMELDna and calcMELDThree functions to
calculate the different MELD scores for each row entry in the provided
dataframe. The provided dataframe should correspond to the dataframe created
by running the prepareData function on the raw SRTR stathiist_liin datafile
in csv form. The input dataframe needs to have the following columns:
Patient_ID, Data, Albumin, Bilirubin, INR, Serum_Creatinine, Serum_Sodium,
Gender, Dialysis_In_Past_Week.
}
\examples{
# Example 1
# Load the sample cleaned up data from the CleanedUpSamplePatientData.rda
# file,  then use it ro run the createMELDdf function
\dontrun{
load("../data/cleanedUpSamplePatientData.rda")
createMELDdf(cleanedUpSamplePatientData)
}

}
\references{
Wickham H, François R, Henry L, Müller K, Vaughan D (2023).
\emph{dplyr: A Grammar of Data Manipulation.} R package version 1.1.4,
https://github.com/tidyverse/dplyr, https://dplyr.tidyverse.org.

Wickham H, Henry L (2023). \emph{purrr: Functional Programming Tools.}
R package version 1.0.2,
https://github.com/tidyverse/purrr, https://purrr.tidyverse.org/.

Wickham H, Hester J, Chang W, Bryan J (2022).
\emph{devtools: Tools to Make Developing R Packages Easier.}
https://devtools.r-lib.org/, https://github.com/r-lib/devtools.

Wickham H, Vaughan D, Girlich M (2023). \emph{tidyr: Tidy Messy Data.}
R package version 1.3.0,
https://github.com/tidyverse/tidyr, https://tidyr.tidyverse.org.
}
