% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Process data and extract relevant biomarkers}
\usage{
prepareData(filePath)
}
\arguments{
\item{file_path}{the csv file path for the data set containing the data. It
should have the the following coloumns at least Date, Sex,
DialysisTwiceInPastWeek, Creatinine, INR, Total_Bilirubin, Sodium, Albumin}
}
\value{
selectedData, a dataframe containing the relevant columns to
calculate MELD scores
}
\description{
process the clincial serum blood test and relevant data from a csv data set
and return a dataframe containing only coloumns relevant to the calculation
of MELD scores
}
\examples{
#Example 1
prepareDate(../data/liver_data)

}
\references{
UCI Machine Learning. (2017). Indian Liver Patient Records \link{Data set}.
}
