% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLivTransplantEvolution.R
\name{runLivTransplantEvolution}
\alias{runLivTransplantEvolution}
\title{Launch Shiny App for LivTransplantEvolution}
\usage{
runLivTransplantEvolution()
}
\value{
No return value, but opens up a Shiny page.
}
\description{
A function that launches the Shiny app for LivTransplantEvolution.
This app is designed to facilitate the analysis of liver transplant
relevant risk scores and biomarkers, using the data and functionalities
provided by the LivTransplantEvolution package. The code for the Shiny
app is located in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

LivTransplantEvolution::runLivTransplantEvolution()
}


}
