% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleMELDdf}
\alias{sampleMELDdf}
\title{Sample meld data frame containing MELD score data generated by the
createMELDdf function and made for users to use as input to the visualizeMELD
function to test the visualization}
\format{
A dataframe with columns:
\describe{
\item{Patient_ID}{Contains the numeric patient ID of the patient the data in entries
in that row are for}
\item{Date}{The date that corresponding to the data in that row}
\item{MELD}{MELD score calculated from biomarkers values matching the data
and patient_id in original raw data using the calcMELD function in the
LivTransplantEvolution package}
\item{MELDna}{MELD-na score calculated from biomarkers values matching the data
and patient_id in original raw data using the calcMELDna function in the
LivTransplantEvolution package}
\item{MELD3_0}{MELD-3.0 score calculated from biomarkers values matching the data
and patient_id in original raw data using the calcMELDThree function in the
LivTransplantEvolution package}
}
}
\source{
Created by Mouaid Alim the package author.
}
\usage{
sampleMELDdf
}
\description{
Sample patient MELD score data generated by calling the createMELDdf function
on the cleanedUpSamplePatientData data frame found in the
data/cleanedUpSamplePatientData.rda. This sample data is ideal data that
would be created from running the prepareData function on the stathiist_liin
data file from the SRTR database, and then running createMELDdf on the output
of that function call. This sample dataframe is ideal to use to test out the
visualizeMELD function
}
\examples{
\dontrun{
 sampleMELDdf
}
}
\keyword{datasets}
