% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMELDThree.R
\name{calcMELDThree}
\alias{calcMELDThree}
\title{Calculate the Model for End-Stage Liver Disease-3.0 (MELD-3.0) risk score}
\usage{
calcMELDThree(
  sex = "M",
  creatinine = 1.3,
  bilirubin = 1.2,
  inr = 1.1,
  sodium = 140,
  albumin = 4
)
}
\arguments{
\item{creatinine}{a numeric value of type 'double' for the creatinine
biomarker which reflects kidney function, as kidney dysfunction is often
associated with worsening Liver function. Unit: mg/dl}

\item{bilirubin}{a numeric value of type 'double' for the bilirubin biomarker
that measures the ability of the liver to process byproducts from the
breakdown of old red blood cells. Elevated levels in the blood
(hyperbilirubinemia) indicate poor liver function. Unit: mg/dl}

\item{inr}{a numeric value of type 'double' for International Normalized
Ratio (INR). INR is a blood clotting test that is a measure how quickly your
blood clots and basically how thin your blood is. It is unitless}

\item{sodium}{a numeric value of type 'double' for the serum sodium level.
Another indicator of Kidney function. Sodium is only taken into account for
the MELD-na calculation if the MELD is > 11. Unit: mEq/L}

\item{albumin}{a numeric value of type 'double' for the albumin biomarker.
Albumin is a liver synthesized protein which low levels of can indicate
chronic liver disease or liver dysfunction. Unit: mg/dL}

\item{Sex}{a boolean value indicating whether the patient is female or not.
As females get extra points in the calculation to account for biases.}
}
\value{
meldThreeScore The 'double' value returned from calculating the
MELD-3.0 score calculation based on the paper by Kim et al.
}
\description{
A function that calculates the MELD-3.0 score using the formula
introduced by Kim et all.
Model for End-Stage Liver Disease-3.0 (MELD-3.0) risk score is used to
measure theseverity of Liver dysfunction. It indicates the estimated 90-day
chances of survival for this patient. MELD-3.0 was created as an attempt to
overcome some of the equity issues associated with the MELD and MELD-na
scores. The calculation of MELD-3.0 requires 5 biomarkers, the
patient's Creatinine, Bilirubin and International Normalized Ratio (INR),
Albumin and Serum Sodium levels. The MELD-3.0 informs the estimated 90 day
chances of survival of that patient and can be interpreted using this
formula: 0.946^((0.17698 \* MELD 3.0) – 3.56) \* 100
}
\examples{
#Example 1
# Calculate the MELD-3.0 score for a female

calcMELDThree('F', 2.3, 1.9, 2, 145, 2)

}
\references{
Kim WR, Mannalithara A, Heimbach JK, Kamath PS, Asrani SK, Biggins SW,
Wood NL, Gentry SE, Kwong AJ. MELD 3.0: The Model for End-Stage Liver Disease
Updated for the Modern Era. Gastroenterology. 2021 Dec;161(6):1887-1895.e4.
doi: 10.1053/j.gastro.2021.08.050. Epub 2021 Sep 3. PMID: 34481845;
PMCID: PMC8608337.

Kamath PS, Wiesner RH, Malinchoc M, Kremers W, Therneau TM, Kosberg CL,
D'Amico G, Dickson ER, Kim WR. A model to predict survival in patients with
end-stage liver disease. Hepatology. 2001 Feb;33(2):464-70.
doi: 10.1053/jhep.2001.22172. PMID: 11172350.
}
