% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMELDdf.R
\name{createMELDdf}
\alias{createMELDdf}
\title{Create a data frame with all MELD scores}
\usage{
createMELDdf(data)
}
\arguments{
\item{data}{dataframe created by the preapreData function}
}
\value{
MELDdf, a dataframe containing the different values of the differnet
MELD scores by date
}
\description{
This function uses the functions to calculate the different MELD scores to
to then create a dataframe containing MELD score values corresopnding to each
row in the dataframe created by the preapreData function
}
\examples{
#Example 1
createMELDdf(liver_df)

}
