% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeMELD.R
\name{visualizeMELD}
\alias{visualizeMELD}
\title{Function to plot MELD scores}
\usage{
visualizeMELD(meldScoresdf)
}
\arguments{
\item{meldScoresdf}{a data frame containing the MELD scores organized by date
for MELD, MELDna and MELD-3.0 generated using the createMELDdf function}
}
\value{
no return, plots MELD scores using ggplot2
}
\description{
Visualize all the MELD scores for the patietns
}
\examples{
#Example 1
visualizeMELD(MELD_scores_df)

}
\references{
Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4,
https://ggplot2.tidyverse.org.

Wickham H, Vaughan D, Girlich M (2023). tidyr: Tidy Messy Data. R package
version 1.3.0, https://github.com/tidyverse/tidyr,
https://tidyr.tidyverse.org.
}
